requirements.txt
gradio
tensorflow
Pillow
numpy